requirements.txt
asgiref==3.8.1
Django==5.1.3
mysql-connector-python==9.1.0
PyMySQL==1.1.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
